backend/requirements.txt
Django==2.1.1
django-cors-headers==2.4.0
django-debug-toolbar==1.10.1
django-pyodbc-azure==2.1.0.0
django-rest-registration==0.3.12
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
mysqlclient==1.3.13
Pillow==5.2.0
PyJWT==1.6.4
pyodbc==4.0.24
pytz==2018.5
sqlparse==0.2.4
